chore: update by Sam's code review
