+++ flashinfer/aot.py
@@ -88,6 +88,7 @@ from .jit.mamba import (
     gen_selective_state_update_module,
     gen_selective_state_update_sm90_module,
 )
+from .jit.mhc import gen_mhc_module
 from .jit.mla import gen_mla_module
 from .jit.api_log_stats import gen_api_log_stats_module
 from .jit.norm import gen_norm_module
@@ -573,6 +574,7 @@ def gen_all_modules(
         jit_specs += [
             gen_api_log_stats_module(),
             gen_cascade_module(),
+            gen_mhc_module(),
             gen_norm_module(),
             gen_page_module(),
             gen_quantization_module(),
